refactoring: api and other
