+++ package.json
@@ -15,35 +15,32 @@
   },
   "dependencies": {
     "@quasar/extras": "^1.16.12",
-    "jose": "^5.6.3",
-    "pinia": "^2.1.7",
-    "quasar": "^2.16.5",
-    "vue": "^3.4.29",
-    "vue-router": "^4.4.0"
+    "jose": "^5.7.0",
+    "pinia": "^2.2.2",
+    "quasar": "^2.16.9",
+    "vue": "^3.4.38",
+    "vue-router": "^4.4.3"
   },
   "devDependencies": {
-    "@eslint/js": "^9.8.0",
+    "@eslint/js": "^9.9.0",
     "@quasar/vite-plugin": "^1.7.0",
-    "@rushstack/eslint-patch": "^1.8.0",
     "@tsconfig/node20": "^20.1.4",
     "@types/jsdom": "^21.1.7",
     "@types/node": "^20.14.5",
-    "@vitejs/plugin-vue": "^5.0.5",
-    "@vue/eslint-config-prettier": "^9.0.0",
-    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vitejs/plugin-vue": "^5.1.2",
     "@vue/test-utils": "^2.4.6",
     "@vue/tsconfig": "^0.5.1",
     "eslint": "8.57.0",
     "eslint-plugin-vue": "^9.27.0",
-    "globals": "^15.8.0",
-    "jsdom": "^24.1.0",
+    "globals": "^15.9.0",
+    "jsdom": "^24.1.1",
     "npm-run-all2": "^6.2.0",
-    "prettier": "^3.2.5",
-    "sass": "^1.33.0",
-    "typescript": "~5.4.0",
-    "typescript-eslint": "^7.18.0",
-    "vite": "^5.3.1",
+    "prettier": "^3.3.3",
+    "sass": "^1.77.8",
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^7.3.1",
+    "vite": "^5.4.2",
     "vitest": "^1.6.0",
-    "vue-tsc": "^2.0.21"
+    "vue-tsc": "^2.0.29"
   }
 }
