+++ src/entities/license/api/getDigestFIle.ts
@@ -1,7 +1,8 @@
 import { BACKEND_CONNECTION } from '@/shared/config'
 import { Cookies } from 'quasar'
+import type { LicenseFile } from '../model'
 
-export const getDigestFile = async (id: string) =>
+export const getDigestFile = async (id: string): Promise<LicenseFile> =>
   await fetch(BACKEND_CONNECTION + 'machine_digest_file/' + id, {
     headers: {
       Authorization: 'Bearer ' + Cookies.get('session'),
@@ -9,9 +10,7 @@ export const getDigestFile = async (id: string) =>
   })
     .then((r) => (r.ok ? r : Promise.reject(r.status)))
     .then(async (r) => ({
-      filename: r.headers
-        .get('Content-Disposition')
-        ?.match(/filename="([^"]+)"/)?.[1],
+      filename: r.headers.get('Content-Disposition')?.match(/filename="([^"]+)"/)?.[1],
       blob: await r.blob(),
     }))
     .catch((e) => Promise.reject(e))
