+++ src/entities/license/model/License.ts
