+++ src/entities/license/model/LicenseDto.ts
@@ -1,0 +1,9 @@
+export type LicenseDto = {
+  company_name: string
+  exp_time: string
+  id: number
+  lic_file_name: string
+  lic_num: string
+  machine_digest_file: string
+  product_name: string
+}
