+++ src/entities/license/model/LicenseFile.ts
@@ -1,0 +1,4 @@
+export type LicenseFile = {
+  filename?: string | null
+  blob: Blob
+}
