+++ src/entities/license/model/NewLicenseDto.ts
@@ -1,0 +1,6 @@
+export type NewLicenseDto = {
+  company_name: string
+  product_name: string
+  license_users_count: number
+  exp_time: string
+}
