+++ src/entities/license/model/index.ts
@@ -1,0 +1,4 @@
+export { type License } from './License'
+export { type NewLicenseDto } from './NewLicenseDto'
+export { type LicenseDto } from './LicenseDto'
+export { type LicenseFile } from './LicenseFile'
