+++ src/entities/role/api/index.ts
@@ -1,5 +1,1 @@
-import { getAllRoles } from './getAllRoles'
-import { createRole } from './createRole'
-import { deleteRole } from './deleteRole'
-import { patchRole } from './patchRole'
-export { getAllRoles, createRole, deleteRole, patchRole }
+export { getAllRoles, createRole, deleteRole, patchRole } from './crudRole'
