+++ src/entities/role/index.ts
@@ -1,2 +1,2 @@
-export type { Role } from './Role'
+export type { Role, RoleDto } from './model'
 export { getAllRoles, createRole, deleteRole, patchRole } from './api'
