+++ src/entities/role/model/NewRoleDto.ts
@@ -1,0 +1,3 @@
+export type NewRoleDTO = {
+  name: string
+}
