+++ src/entities/role/model/index.ts
@@ -1,0 +1,3 @@
+export { type NewRoleDTO } from './NewRoleDto'
+export { type RoleDto, RoleDtoToRole } from './RoleDto'
+export { type Role } from './Role'
