+++ src/entities/user/api/index.ts
@@ -1,17 +1,2 @@
-import { getToken } from './getToken'
-import type { Token } from './getToken'
-import { getUserInfo } from './getUserInfo'
-import { getAllUsers } from './getAllUsers'
-import { createUser } from './createUser'
-import { deleteUser } from './deleteUser'
-import { patchUser } from './patchUser'
-
-export {
-  getToken,
-  type Token,
-  getUserInfo,
-  getAllUsers,
-  createUser,
-  deleteUser,
-  patchUser,
-}
+export { type Token, getToken } from './getToken'
+export { getUserInfo, getAllUsers, createUser, deleteUser, patchUser } from './crudUser'
