+++ src/entities/user/model/NewUserDto.ts
@@ -1,0 +1,3 @@
+export type NewUserDto = {
+  username: string
+}
