+++ src/entities/user/model/User.ts
