+++ src/entities/user/model/index.ts
@@ -1,1 +1,3 @@
-export { useUserStore, type UserInfo } from "./store"
+export { useUserStore, type UserInfo } from './store'
+export type { NewUserDto } from './NewUserDto'
+export type { User } from './User'
