+++ src/features/generateLicense/GenerateLicense.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { generateLicense, type LicenseInfo } from '@/entities/license'
+import { generateLicense, type NewLicenseDto } from '@/entities/license'
 import { exportFile, date, type QFile } from 'quasar'
 import { ref, type Ref } from 'vue'
 import { getErrorByCode, showError } from '../showError'
@@ -21,22 +21,17 @@ const handlePopupToggle = () => {
 }
 
 const handleSubmit = async () => {
+  if (!file.value) return
   loading.value = true
   generateLicense(file.value, {
     company_name: company.value,
     product_name: product.value,
     license_users_count: userCount.value,
-    exp_time: date.formatDate(
-      date.extractDate(expirationDate.value, 'D.M.YYYY'),
-      'DD-MM-YYYY'
-    ),
-  } as LicenseInfo)
+    exp_time: date.formatDate(date.extractDate(expirationDate.value, 'D.M.YYYY'), 'DD-MM-YYYY'),
+  } as NewLicenseDto)
     .then((downloadedFile) => {
       emits('AddLicense')
-      exportFile(
-        downloadedFile.filename ?? 'license file.txt',
-        downloadedFile.blob
-      )
+      exportFile(downloadedFile.filename ?? 'license file.txt', downloadedFile.blob)
     })
     .catch((e) => showError(getErrorByCode(e)))
 
@@ -45,21 +40,15 @@ const handleSubmit = async () => {
   product.value = ''
   company.value = ''
   userCount.value = 1
-  expirationDate.value = date.formatDate(
-    date.addToDate(new Date(), { months: 1 }),
-    'D.M.YYYY'
-  )
+  expirationDate.value = date.formatDate(date.addToDate(new Date(), { months: 1 }), 'D.M.YYYY')
 }
 
 const ruLocale = {
-  days: 'Воскресенье_Понедельник_Вторник_Среда_Четверг_Пятница_Суббота_Воскресенье'.split(
+  days: 'Воскресенье_Понедельник_Вторник_Среда_Четверг_Пятница_Суббота_Воскресенье'.split('_'),
+  daysShort: 'Вс_Пн_Вт_Ср_Чт_Пт_Сб'.split('_'),
+  months: 'Январь_Февраль_Март_Апрель_Май_Июнь_Июль_Август_Сентябрь_Октябрь_Ноябрь_Декабрь'.split(
     '_'
   ),
-  daysShort: 'Вс_Пн_Вт_Ср_Чт_Пт_Сб'.split('_'),
-  months:
-    'Январь_Февраль_Март_Апрель_Май_Июнь_Июль_Август_Сентябрь_Октябрь_Ноябрь_Декабрь'.split(
-      '_'
-    ),
   monthsShort: 'Янв_Фев_Мар_Апр_Май_Июн_Июл_Авг_Сен_Окт_Ноя_Дек'.split('_'),
 }
 
@@ -99,18 +88,14 @@ const uploadFile = () => {
         v-model="userCount"
         no-error-icon
         hide-bottom-space
-        :rules="[
-          (v: string) => !!v,
-          (v: string | number) => !isNaN(v as number),
-        ]"
+        :rules="[(v: string) => !!v, (v: string | number) => !isNaN(v as number)]"
       />
       <label>Дата окончания</label>
       <q-input
         v-model="expirationDate"
         :rules="[
           (v: string) => /^[0-3]?\d\.[0-1]?\d\.[\d]+$/.test(v),
-          (v: string) =>
-            date.extractDate(v, 'D.M.YYYY').getTime() > new Date().getTime(),
+          (v: string) => date.extractDate(v, 'D.M.YYYY').getTime() > new Date().getTime(),
         ]"
         placeholder="Введите дату окончания"
         outlined
@@ -119,16 +104,8 @@ const uploadFile = () => {
         no-error-icon
       >
         <template v-slot:append>
-          <q-icon
-            size="14px"
-            name="img:src/shared/assets/calendar.svg"
-            class="cursor-pointer"
-          >
-            <q-popup-proxy
-              cover
-              transition-show="scale"
-              transition-hide="scale"
-            >
+          <q-icon size="14px" name="img:src/shared/assets/calendar.svg" class="cursor-pointer">
+            <q-popup-proxy cover transition-show="scale" transition-hide="scale">
               <q-date
                 v-model="expirationDate"
                 mask="D.M.YYYY"
@@ -138,12 +115,7 @@ const uploadFile = () => {
                 :locale="ruLocale"
               >
                 <div class="row items-center justify-end">
-                  <q-btn
-                    v-close-popup
-                    label="Закрыть"
-                    class="btn btn-fill"
-                    flat
-                  />
+                  <q-btn v-close-popup label="Закрыть" class="btn btn-fill" flat />
                 </div>
               </q-date>
             </q-popup-proxy>
