+++ src/pages/adminPage/AdminPage.vue
@@ -1,19 +1,15 @@
 <script setup lang="ts">
-import { getAllAccesses } from '@/entities/accsses'
-import type { Role } from '@/entities/role'
-import { deleteRole, getAllRoles, patchRole } from '@/entities/role'
-import { deleteUser, getAllUsers } from '@/entities/user'
-import type { User } from '@/entities/user/User'
-import { Header } from '@/widgets/header'
 import type { Column } from '@/shared/model'
-import { onBeforeMount, ref } from 'vue'
+import { getAllAccesses, type Access } from '@/entities/accsses'
+import { type Role, deleteRole, getAllRoles, patchRole } from '@/entities/role'
+import { type User, patchUser, deleteUser, getAllUsers } from '@/entities/user'
 import { getErrorByCode, showError } from '@/features/showError'
-import { useQuasar } from 'quasar'
 import { CreateRole } from '@/features/createRole'
 import { CreateUser } from '@/features/createUser'
 import { RoleAssignmentList } from '@/widgets/roleAssignmentList'
-import { patchUser } from '@/entities/user/api'
-import { RoleDtoToRole } from '@/entities/role/api/getAllRoles'
+import { Header } from '@/widgets/header'
+import { onBeforeMount, ref } from 'vue'
+import { useQuasar } from 'quasar'
 
 const $q = useQuasar()
 const tab = ref<'users' | 'roles'>('users')
@@ -67,8 +63,8 @@ const handleUpdate = () => {
 
   Promise.all([rolesPromise, accessesPromise, usersPromise])
     .then((r) => {
-      roles.value = r[0].map((i) => RoleDtoToRole(i))
-      authoritiesSelectOptions.value = r[1].map((v) => v.name)
+      roles.value = r[0]
+      authoritiesSelectOptions.value = r[1].map((v: Access) => v.name)
       usersData.value = r[2]
       roleSelectOptions.value = roles.value.map((v) => v.name)
     })
@@ -76,15 +72,11 @@ const handleUpdate = () => {
 }
 
 const handleDeleteRole = async (role: Role[]) => {
-  await Promise.all(role.map((r) => deleteRole(r))).catch((e) =>
-    showError(getErrorByCode(e))
-  )
+  await Promise.all(role.map((r) => deleteRole(r))).catch((e) => showError(getErrorByCode(e)))
   handleUpdate()
 }
 const handleDeleteUser = async (user: User[]) => {
-  await Promise.all(user.map((u) => deleteUser(u))).catch((e) =>
-    showError(getErrorByCode(e))
-  )
+  await Promise.all(user.map((u) => deleteUser(u))).catch((e) => showError(getErrorByCode(e)))
   handleUpdate()
 }
 
@@ -107,18 +99,8 @@ onBeforeMount(() => {
     <q-page>
       <q-card flat class="padded q-mx-md q-mb-md card">
         <q-tabs align="left" narrow-indicator dense v-model="tab">
-          <q-tab
-            :ripple="false"
-            class="menu-item text-body1"
-            name="users"
-            label="Пользователи"
-          />
-          <q-tab
-            :ripple="false"
-            class="menu-item text-body1"
-            name="roles"
-            label="Роли"
-          />
+          <q-tab :ripple="false" class="menu-item text-body1" name="users" label="Пользователи" />
+          <q-tab :ripple="false" class="menu-item text-body1" name="roles" label="Роли" />
         </q-tabs>
         <q-separator />
         <q-tab-panels v-model="tab" keep-alive>
